rename nav to Nav
